requirements.txt
pyzmq~=25.0.0
pydicom~=2.3.1